+++ GoalMogul-iOS/CHANGELOG.md
@@ -10,6 +10,8 @@ if there is corresponding issue, state it
 
 * Version 0.0.1
 ## 1 June 2018, by Jia Zeng <andyzeng96@gmail.com>
+* remove friends … button
+* change profileSummaryCard button to blue background circle >
 * Empty results for search results (No Results)
 * change search card button to blue background circle >
 * A lot of refactoring